Interpolation - 1
Interpolation- To generate or view the dynamic data on the front either from a function or from an api call then interpolation is used. in the .ts file and viewed in the component.html file.

We use the curly braces and name of the property to view the dynamic data {{propertyname}}

Just to keep in mind we can not increment a number in the propertybeing called and cannot use the new keyword as well. We can do addition subtraction.